Move DWI specific functions from filemanip utils to dwi utils preproc pipeline utils
