+++ test/unittests/pipelines/dwi/preprocessing/test_common_preprocessing_utils.py
@@ -3,6 +3,8 @@ import numpy as np
 import pytest
 from numpy.testing import assert_array_equal
 
+from clinica.utils.exceptions import ClinicaException
+
 
 @pytest.mark.parametrize("phase", ["a", "X", "x+", "-x", "i", "j", "k", "foo"])
 def test_generate_acq_file_errors(tmp_path, phase):
@@ -177,3 +179,120 @@ def test_check_dwi_volume(tmp_path, dwi_dataset):
     from clinica.utils.testing_utils import build_dwi_dataset
 
     check_dwi_volume(build_dwi_dataset(tmp_path, 9, 9, 9))
+
+
+@pytest.mark.parametrize(
+    "dictionary, expected",
+    [
+        (
+            {
+                "TotalReadoutTime": 1,
+                "EstimatedTotalReadoutTime": "",
+                "PhaseEncodingSteps": "",
+                "PixelBandwidth": "",
+                "PhaseEncodingDirection": "j+",
+                "PhaseEncodingAxis": "",
+            },
+            (1, "y"),
+        ),
+        (
+            {
+                "EstimatedTotalReadoutTime": 1,
+                "PhaseEncodingAxis": "j",
+            },
+            (1, "y"),
+        ),
+        (
+            {
+                "PhaseEncodingSteps": 1,
+                "PixelBandwidth": 0.5,
+                "PhaseEncodingDirection": "j-",
+            },
+            (2, "y-"),
+        ),
+    ],
+)
+def test_get_readout_time_and_phase_encoding_direction(tmp_path, dictionary, expected):
+    """This function tests that the outputs of `extract_metadata_from_json` are what you'd expect in the case of DWI."""
+    import json
+
+    from clinica.pipelines.dwi.preprocessing.utils import (
+        get_readout_time_and_phase_encoding_direction,
+    )
+
+    with open(tmp_path / "metadata.json", "w") as outfile:
+        json.dump(dictionary, outfile)
+
+    assert (
+        get_readout_time_and_phase_encoding_direction(tmp_path / "metadata.json")
+        == expected
+    )
+
+
+@pytest.mark.parametrize(
+    "input_list, dictionary, error_type, error_log",
+    [
+        (
+            [
+                "TotalReadoutTime",
+                "PhaseEncodingDirection",
+            ],
+            {
+                "PhaseEncodingSteps": 1,
+                "PixelBandwidth": 0.5,
+            },
+            ClinicaException,
+            "Could not recover the PhaseEncodingDirection from JSON file.",
+        ),
+        (
+            [
+                "TotalReadoutTime",
+                "PhaseEncodingDirection",
+            ],
+            {
+                "PhaseEncodingDirection": "j+",
+            },
+            ClinicaException,
+            "Could not recover the TotalReadoutTime from JSON file.",
+        ),
+        (
+            [
+                "TotalReadoutTime",
+                "PhaseEncodingDirection",
+            ],
+            {
+                "PhaseEncodingSteps": 1,
+                "PixelBandwidth": 0,
+                "PhaseEncodingDirection": "j+",
+            },
+            ValueError,
+            "Pixel Bandwidth value is not valid.",
+        ),
+        (
+            [
+                "blabla",
+                "PhaseEncodingDirection",
+            ],
+            {
+                "PhaseEncodingDirection": "j+",
+            },
+            ClinicaException,
+            "Could not recover the TotalReadoutTime from JSON file.",
+        ),
+    ],
+)
+def test_get_readout_time_and_phase_encoding_direction_errors(
+    tmp_path, input_list, dictionary, error_type, error_log
+):
+    """This function tests that `extract_metadata_from_json` errors as expected in the case of DWI."""
+    import json
+
+    from clinica.pipelines.dwi.preprocessing.utils import (
+        get_readout_time_and_phase_encoding_direction,
+    )
+
+    with open(tmp_path / "metadata.json", "w") as outfile:
+        json.dump(dictionary, outfile)
+
+    with pytest.raises(error_type, match=error_log):
+        get_readout_time_and_phase_encoding_direction(tmp_path / "metadata.json")
